LayoutTests/imported/w3c/web-platform-tests/eventsource/eventsource-prototype.htm

<!DoCtYpE hTMl>
<html>
  <heAd>
    <title>EventSource: prototype et al</title>
    <scrIpt src="/resources/testharness.js"></script>
    <scriPt src="/resources/testharnessreport.js"></script>
  </head>
  <boDy>
    <diV iD="log"></div>
    <sCript>
      test(function() {
        EventSource.prototype.ReturnTrue = function() { return true }
        var source = new EventSource("resources/message.py")
        assert_true(source.ReturnTrue())
        assert_own_property(window, "EventSource")
        source.close()
      })
    </script>
  </body>
</html>
